Fix circular imports related to XLabsControlNetFluxOutput and InstantXControlNetFluxOutput.
